+++ lib/src/accounting/server_state.rs
@@ -4,10 +4,9 @@ use anyhow::Context;
 use chrono::{DateTime, FixedOffset};
 use lrzcc_wire::accounting::{
     ServerState, ServerStateCreateData, ServerStateImport,
-    ServerStateModifyData,
+    ServerStateListParams, ServerStateModifyData,
 };
 use reqwest::blocking::Client;
-use reqwest::Url;
 use reqwest::{Method, StatusCode};
 use std::rc::Rc;
 
@@ -21,10 +20,7 @@ pub struct ServerStateListRequest {
     url: String,
     client: Rc<Client>,
 
-    server: Option<String>,
-    user: Option<u32>,
-    project: Option<u32>,
-    all: bool,
+    params: ServerStateListParams,
 }
 
 impl ServerStateListRequest {
@@ -32,30 +28,24 @@ impl ServerStateListRequest {
         Self {
             url: url.to_string(),
             client: Rc::clone(client),
-            server: None,
-            user: None,
-            project: None,
-            all: false,
-        }
-    }
 
-    fn params(&self) -> Vec<(&str, String)> {
-        let mut params = Vec::new();
-        if let Some(server) = &self.server {
-            params.push(("server", server.to_string()));
-        } else if let Some(user) = self.user {
-            params.push(("user", user.to_string()));
-        } else if let Some(project) = self.project {
-            params.push(("project", project.to_string()));
-        } else if self.all {
-            params.push(("all", "1".to_string()));
+            params: ServerStateListParams {
+                server: None,
+                user: None,
+                project: None,
+                all: None,
+            },
         }
-        params
     }
 
     pub fn send(&self) -> Result<Vec<ServerState>, ApiError> {
-        let url = Url::parse_with_params(self.url.as_str(), self.params())
-            .context("Could not parse URL GET parameters.")?;
+        let params = serde_urlencoded::to_string(&self.params)
+            .context("Failed to envode URL parameters")?;
+        let url = if params.is_empty() {
+            self.url.clone()
+        } else {
+            format!("{}?{}", self.url, params)
+        };
         request(
             &self.client,
             Method::GET,
@@ -66,22 +56,22 @@ impl ServerStateListRequest {
     }
 
     pub fn server(&mut self, server: &str) -> &mut Self {
-        self.server = Some(server.to_string());
+        self.params.server = Some(server.to_string());
         self
     }
 
     pub fn user(&mut self, user: u32) -> &mut Self {
-        self.user = Some(user);
+        self.params.user = Some(user);
         self
     }
 
     pub fn project(&mut self, project: u32) -> &mut Self {
-        self.project = Some(project);
+        self.params.project = Some(project);
         self
     }
 
     pub fn all(&mut self) -> &mut Self {
-        self.all = true;
+        self.params.all = Some(true);
         self
     }
 }