+++ wire/src/budgeting/user_budget.rs
@@ -20,6 +20,14 @@ impl Display for UserBudget {
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct UserBudgetListParams {
+    pub user: Option<u32>,
+    pub project: Option<u32>,
+    pub all: Option<bool>,
+    pub year: Option<u32>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct UserBudgetCreateData {
     pub user: u32,
     #[serde(skip_serializing_if = "Option::is_none")]